cat '' 에러수정, 서브쉘 작동 확인, libft 추가

+++ srcs/utils/general_utils.c
@@ -63,8 +63,8 @@ char	*modified_substr(char *s, int start, int len)
 
 char	*modified_strtrim(char const *s1, char const *set)
 {
-	size_t	start;
-	size_t	end;
+	int		start;
+	int		end;
 	char	*str;
 
 	if (set == 0)
